backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-dotenv==1.0.0
boto3==1.34.0
pymongo==4.6.0
motor==3.3.2

